Move ToolSearchToolCallingAdvisor to spring-ai-tool-search-advisor
This commit moves TOOL_SEARCH_TOOL_SESSION_ID_KEY from ToolSearchToolCallingAdvisor
to ToolSearchTool to avoid a circular dependency issue.

Its package also changes from org.springframework.ai.tool.toolsearch.advisor to
org.springframework.ai.chat.client.advisor.toolsearch.

Closes #6309
