+++ advisors/spring-ai-tool-search-advisor/src/main/java/org/springframework/ai/chat/client/advisor/toolsearch/package-info.java
@@ -18,6 +18,6 @@
  * Tool Search Tool Advisor package for dynamic tool discovery and search capabilities.
  */
 @NullMarked
-package org.springframework.ai.tool.toolsearch.advisor;
+package org.springframework.ai.chat.client.advisor.toolsearch;
 
 import org.jspecify.annotations.NullMarked;
